+++ src/Forms/FormDecorator.php
@@ -26,7 +26,7 @@ class FormDecorator extends ModelAdminDecorator
 	protected function modifyListingQuery(Builder $query)
 	{
 		$query->orderBy($this->model->getTable() . '.name')->with(['submissions' => function($query) {
-			$query->selectRaw('count(*) as count, form_id')->groupBy('form_id');
+			$query->selectRaw('count(*) as count, max(created_at) as last_submission, form_id')->groupBy('form_id');
 		}]);
 	}
 
@@ -43,6 +43,8 @@ class FormDecorator extends ModelAdminDecorator
 			'Newsletter Signup' => $instance->newsletter_signup ? '<i class="fa fa-check"></i>' : '',
 			'Status' => $instance->getAttribute('status') == 1 ? '<i class="fa fa-check"></i>' : '',
 			'Submissions' => $instance->submissions->first()->count,
+			'Last Submission' => $instance->submissions->first()->last_submission ?
+				date('d M Y - H:i', strtotime($instance->submissions->first()->last_submission)) : '-',
 		];
 	}